Major restructuring of the block with tests.
